requirements.txt
flask-restplus==0.10.1
Werkzeug==0.16.1
Flask-Cors==3.0.10
googlemaps==4.4.5
dicttoxml==1.7.4
python-dotenv==0.15.0
